[move] 폴더 구조 이동 #16
