+++ src/main/java/org/example/fanzip/payment/controller/PaymentsController.java
@@ -1,8 +1,8 @@
-package org.example.fanzip.controller;
+package org.example.fanzip.payment.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.example.fanzip.dto.PaymentsDto;
-import org.example.fanzip.service.PaymentService;
+import org.example.fanzip.payment.dto.PaymentsDto;
+import org.example.fanzip.payment.service.PaymentService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,5 +17,4 @@ public class PaymentsController {
         PaymentsDto created = paymentService.createdPayment(paymentsDto);
         return ResponseEntity.ok(created);
     }
-
 }
