+++ src/main/java/org/example/fanzip/payment/service/PaymentService.java
@@ -1,6 +1,6 @@
-package org.example.fanzip.service;
+package org.example.fanzip.payment.service;
 
-import org.example.fanzip.dto.PaymentsDto;
+import org.example.fanzip.payment.dto.PaymentsDto;
 
 public interface PaymentService {
     PaymentsDto createdPayment(PaymentsDto paymentsDto);
