+++ src/main/resources/mybatis-config.xml
@@ -4,6 +4,6 @@
         "http://mybatis.org/dtd/mybatis-3-config.dtd">
 <configuration>
     <typeAliases>
-        <typeAlias alias="PaymentsDto" type="org.example.fanzip.dto.PaymentsDto"></typeAlias>
+        <typeAlias alias="PaymentsDto" type="org.example.fanzip.payment.dto.PaymentsDto"></typeAlias>
     </typeAliases>
 </configuration>
